Rename collapse_edges
